Update vector.py
This fix helps the guide building correct joint axis when the given points are collinears (and so the two resulting vectors for plane normal solving are parallels). 
Next step for more precision : we could add an extra parameter to this function where we gie up axis of the guide, so up axis of the guide will be the same as up axis of the chain, and only when vectors are parallels.